Integer vertex attributes

+++ src/main/resources/assets/canvas/shaders/internal/vertex.glsl
@@ -8,8 +8,8 @@
 in vec3 in_vertex;
 in vec4 in_color;
 in vec2 in_uv;
-in float in_sprite;
-in float in_material;
+in uint in_sprite;
+in uint in_material;
 	#ifdef VANILLA_LIGHTING
 in vec4 in_lightmap;
 in vec4 in_normal_flags;
